Move Sentry initialization to utility module
